+++ src/extension.ts
@@ -1,34 +1,134 @@
 import * as vscode from 'vscode';
+import * as odoo_lib from 'odoo-xmlrpc';
+
+
+function get_odoo_class() {
+	// Get class name
+	var editor = vscode.window.activeTextEditor;
+	if (!editor) {
+		vscode.window.showInformationMessage('No editor');
+		return ''; // No open text editor
+		}
+	var selection = editor.selection;
+	if (selection.isEmpty === true) {
+		vscode.window.showErrorMessage('Select an odoo class _name or _inherit');
+		return '';
+		}
+	var text = editor.document.getText(selection);
+	var regex = /(|.+)(_name|_inherit)(|.+)=(|.+)("|')(.+)("|')/g;
+	var regexp = new RegExp(regex);
+	var test_regexp = regexp.test(text);
+	if (!test_regexp) {
+		vscode.window.showErrorMessage('Select an odoo class _name or _inherit');
+		return '';
+		}
+	var match_regex = regex.exec(text);
+	if (!match_regex) {
+		return '';
+		}
+	var odoo_class = match_regex[6];
+	return odoo_class;
+}
+
+
+function create_model_table(odoo_class: string, odoo: any) {
+	var inParams = [];
+	inParams.push([['model', '=', odoo_class]]);
+	inParams.push(['id', 'model', 'field_id']);
+	inParams.push(0); // offset
+	inParams.push(5); // limit
+	var params = [];
+	params.push(inParams);
+	odoo.execute_kw('ir.model', 'search_read', params,
+		function (err: any, value: any) {
+			if (err) {
+				vscode.window.showErrorMessage('Error: ' + err);
+				return;
+				}
+		if (value.length === 0) {
+			vscode.window.showErrorMessage('Impossible to find data for ' + odoo_class);
+			}
+		// vscode.window.showInformationMessage(value[0].model);
+		const field_ids = value[0].field_id;
+
+		
+		var inParams = [];
+		inParams.push(field_ids); //ids
+		inParams.push(['id', 'name', 'ttype', 'modules']); //fields
+		var params = [];
+		params.push(inParams);
+		odoo.execute_kw('ir.model.fields', 'read', params, function (err2: any, fields: any) {
+			if (err2) { return; }
+			// Show results
+			const panel = vscode.window.createWebviewPanel(
+				'odoo.dev.model.result.' + odoo_class,
+				'Model:' + odoo_class,
+				vscode.ViewColumn.One, );
+			var content_html = `<!DOCTYPE html>
+				<html lang="en">
+				<head>
+					<meta charset="UTF-8">
+					<meta name="viewport" content="width=device-width, initial-scale=1.0">
+					<title>Cat Coding</title>
+					<style>
+					table {
+						width: 100%;
+					}
+					table, th, td {
+						border: 1px solid white;
+						border-collapse: collapse;
+						padding: 3px;
+						padding: 3px;
+					}
+					td {
+						vertical-align: middle;
+					}
+					</style>
+				</head>
+				<body>
+					<div style="width:100%;">
+						<table>
+							<thead>
+								<tr>
+									<th>ID</th>
+									<th>FIELD NAME</th>
+									<th>TYPE</th>
+									<th>IN APPS</th>
+								</tr>
+							</thead>
+							<tbody>`;
+			var i = 0;
+			for (i = 0; i < fields.length; i++) {
+				content_html += '<tr>';
+				content_html += '<td>' + fields[i].id + '</td>';
+				content_html += '<td>' + fields[i].name + '</td>';
+				content_html += '<td>' + fields[i].ttype + '</td>';
+				content_html += '<td>' + fields[i].modules + '</td>';
+				content_html += '</tr>';
+			}
+			content_html += `</tbody>
+						</table>
+					</div>
+				</body>
+				</html>`;
+			panel.webview.html = content_html;
+			});
+		});
+	}
+
 
 export function activate(context: vscode.ExtensionContext) {
 
-	console.log('Congratulations, your extension "vsc-odoo-development" is now active!');
+	console.log('Extension "vsc-odoo-development" is now active!');
 
-	let disposable = vscode.commands.registerCommand('extension.dev.odoo.create.security.rule', () => {
+	let disposable_security_rule = vscode.commands.registerCommand('extension.dev.odoo.create.security.rule', () => {
+		// Get class name
+		var odoo_class_dot = get_odoo_class();
+		if (odoo_class_dot === '') {
+			return;
+			}
 		var editor = vscode.window.activeTextEditor;
-		if (!editor) {
-			vscode.window.showInformationMessage('No editor');
-			return; // No open text editor
-			}
-		var selection = editor.selection;
-		if (selection.isEmpty === true) {
-			vscode.window.showErrorMessage('Select an odoo class _name or _inherit');
-			return;
-			}
-		var text = editor.document.getText(selection);
-		var regex = /(|.+)(_name|_inherit)(|.+)=(|.+)("|')(.+)("|')/g;
-		var regexp = new RegExp(regex);
-		var test_regexp = regexp.test(text);
-		if (!test_regexp) {
-			vscode.window.showErrorMessage('Select an odoo class _name or _inherit');
-			return;
-			}
-		var match_regex = regex.exec(text);
-		console.debug(match_regex);
-		if (!match_regex) {
-			return;
-			}
-		var odoo_class_dot = match_regex[6];
+		if (!editor) {return;}
 		var odoo_class_underscore = odoo_class_dot.split('.').join('_');
 		// Get module name
 		// Presume module strucure is compatible with Odoo/OCA guidelines
@@ -51,7 +151,35 @@ export function activate(context: vscode.ExtensionContext) {
 			'Odoo security rule created in clipboard for ' + odoo_class_dot);
 		});
 
-	context.subscriptions.push(disposable);
+	context.subscriptions.push(disposable_security_rule);
+
+	let disposable_model_remote = vscode.commands.registerCommand('extension.dev.odoo.model.from.remote', () => {
+		// Get class name
+		var odoo_class = get_odoo_class();
+		if (odoo_class === '') {
+			return;
+			}
+		// Open connection with Odoo
+		var odoo_dev_config = vscode.workspace.getConfiguration('odoo.dev');
+		var odoo = new odoo_lib({
+			url: odoo_dev_config.get('OdooRemoteUrl'),
+			port: odoo_dev_config.get('OdooRemotePort'),
+			db: odoo_dev_config.get('OdooRemoteDb'),
+			username: odoo_dev_config.get('OdooRemoteUser'),
+			password: odoo_dev_config.get('OdooRemotePassword')
+			});
+		odoo.connect(function (err: any) {
+			if (err) {
+				vscode.window.showErrorMessage('Impossibile to connect with Odoo');
+				return;
+				}
+			create_model_table(odoo_class, odoo);
+			});
+		});
+
+	context.subscriptions.push(disposable_model_remote);
+
 }
+
 
 export function deactivate() {}