+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change Log
 
-All notable changes to the "vsc-odoo-development" extension will be documented in this file.
+## 0.2.0
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- Show Odoo remote model information from class _name or _inherit
 
-## [Unreleased]
+## 0.1.0
 
+- Create security rule from class _name or _inherit
 - Initial release